fix: fixed FoodItem constructor

+++ FourthPracticalTask/Task_4_2/src/FoodItem.java
@@ -4,7 +4,7 @@ public class FoodItem extends GenericItem {
     Date expirationDate;
     short discount;
 
-    public FoodItem(String name, double price, Category category, Date expirationDate, short discount) {
+    public FoodItem(String name, float price, Category category, Date expirationDate, short discount) {
         super(name, price, category);
         this.expirationDate = expirationDate;
         this.discount = discount;
